some restructuring and refactoring
