+++ FlutterApp/lib/components/meal.dart
@@ -1,0 +1,33 @@
+class Meal {
+  final String? name;
+
+  Meal({this.name});
+}
+
+List<Meal> mealList = [
+  Meal(name: "Brokół"),
+  Meal(name: "Sałatka cezar"),
+  Meal(name: "Marchewka"),
+  Meal(name: "Skrzydełka kurczaka"),
+  Meal(name: "Tort czekoladowy"),
+  Meal(name: "Babeczki"),
+  Meal(name: "Winniczki"),
+  Meal(name: "Frytki"),
+  Meal(name: "Sernik"),
+  Meal(name: "Hamburger"),
+  Meal(name: "Hot dog"),
+  Meal(name: "Lody"),
+  Meal(name: "Lasagne"),
+  Meal(name: "Omlet"),
+  Meal(name: "Naleśniki"),
+  Meal(name: "Pizza"),
+  Meal(name: "Żeberka"),
+  Meal(name: "Jajecznica"),
+  Meal(name: "Zupa"),
+  Meal(name: "Spaghetti bolognese"),
+  Meal(name: "Spaghetti carbonara"),
+  Meal(name: "Stek"),
+  Meal(name: "Sushi"),
+  Meal(name: "Tiramisu"),
+  Meal(name: "Gofry"),
+];
